Move providers into its own domain
