+++ packages/bridge-ui/src/components/Transaction.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
   import type { BridgeTransaction } from '../domain/transactions';
   import { chains, CHAIN_MAINNET, CHAIN_TKO } from '../domain/chain';
-  import { type Chain, providers } from '../domain/chain';
+  import type { Chain } from '../domain/chain';
+  import { providers } from '../domain/provider';
   import { ArrowTopRightOnSquare } from 'svelte-heros-v2';
   import { MessageStatus } from '../domain/message';
   import { Contract, ethers } from 'ethers';
